+++ src/main/java/chess/domain/board/Squares.java
@@ -1,7 +1,9 @@
 package chess.domain.board;
 
-import chess.domain.piece.*;
+import chess.domain.piece.Empty;
+import chess.domain.piece.normal.*;
 import chess.domain.piece.pawn.Pawn;
+import chess.domain.piece.property.Color;
 
 import java.util.ArrayList;
 import java.util.List;
