+++ src/main/java/chess/domain/game/Waiting.java
@@ -3,7 +3,7 @@ package chess.domain.game;
 import chess.domain.board.Board;
 import chess.domain.board.Position;
 import chess.domain.board.Squares;
-import chess.domain.piece.Color;
+import chess.domain.piece.property.Color;
 
 import java.util.List;
