+++ src/main/java/chess/domain/piece/Empty.java
@@ -1,6 +1,9 @@
 package chess.domain.piece;
 
 import chess.domain.board.Position;
+import chess.domain.piece.normal.Piece;
+import chess.domain.piece.property.Color;
+import chess.domain.piece.property.Kind;
 
 import java.util.Map;
 import java.util.Set;
