+++ src/main/java/chess/domain/piece/property/Kind.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.piece.property;
 
 public enum Kind {
     PAWN,
