+++ src/main/java/chess/view/KindMapper.java
@@ -1,8 +1,8 @@
 package chess.view;
 
-import chess.domain.piece.Color;
-import chess.domain.piece.Kind;
-import chess.domain.piece.Piece;
+import chess.domain.piece.normal.Piece;
+import chess.domain.piece.property.Color;
+import chess.domain.piece.property.Kind;
 
 import java.util.Arrays;
 import java.util.List;
