+++ src/test/java/chess/domain/piece/KingTest.java
@@ -3,6 +3,8 @@ package chess.domain.piece;
 import chess.domain.board.File;
 import chess.domain.board.Position;
 import chess.domain.board.Rank;
+import chess.domain.piece.normal.King;
+import chess.domain.piece.property.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
