+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -1,6 +1,8 @@
 package chess.domain.piece;
 
 import chess.domain.board.Position;
+import chess.domain.piece.normal.Knight;
+import chess.domain.piece.property.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
